Add ValidationHelperTest
